api_ingestion/requirements.txt
Flask==2.0.1
requests==2.26.0
google-cloud-pubsub==2.8.0
